+++ examples/enable.rs
@@ -8,11 +8,10 @@
 
 #[macro_use]
 extern crate error_chain;
-extern crate pfctl;
 
 use pfctl::PfCtl;
 
-error_chain!{}
+error_chain! {}
 quick_main!(run);
 
 fn run() -> Result<()> {