+++ examples/transaction.rs
@@ -6,16 +6,13 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-extern crate pfctl;
-
 #[macro_use]
 extern crate error_chain;
 
 use pfctl::PfCtl;
-
 use std::net::Ipv4Addr;
 
-error_chain!{}
+error_chain! {}
 quick_main!(run);
 
 static ANCHOR_NAME: &str = "test.anchor";