+++ src/rule/direction.rs
@@ -6,7 +6,7 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-use ffi;
+use crate::ffi;
 
 /// Enum describing matching of rule towards packet flow direction.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]