+++ tests/helper/pfcli.rs
@@ -1,8 +1,7 @@
-use std::process::Command;
-use std::str;
+use std::{process::Command, str};
 
 mod errors {
-    error_chain!{}
+    error_chain! {}
 }
 use self::errors::*;